requirements.txt
PyChromeDevTools==0.4
psutil==5.9.5
openpyxl==3.1.2
python-dotenv==1.0.0